requirements.txt
Flask==3.0.0
google-generativeai
gTTS
SpeechRecognition
yt-dlp
gunicorn